the new build system
This is the main diff, introducing a straightforward Makefile
and a hand-written ./configure script. This is a replacement
of the GNU autotools monstrosity.

In its current form, this builds the "core" SoX,
i.e. all of the formats and effects that SoX itself implements.
On the other hand, none of the optional external formats
(such as ogg or wavpack) or audio drivers (such as sndio or coreaudio)
gets compiled in, as we don't check for these externals yet.

The rest of the diff is mostly fallout and cleanup:

Get rid of needless home-grown implementations of tmpfile(3)
or strcasesmp(3) etc that are standard on every system now.

Stop checking for unistd.h or fcntl.h etc;
of course we have those, just #include them.

Don't try to accommodate systems that don't have fstat(2) or
fseek(2) or similar functions, standardized for decades by now.

Unify the HAVE_ checks: set each of them to 0 or 1
and check #if HAVE_WHATEVER in the code.

Tested macOS 14.5, more tests coming.
Please test everywhere.