detect and use CoreAudio
This is a simple test of whether the CoreAudio framework is present
(as it is on macOS). This interface is notorious for its lack of
documentation - the test is mostly taken from what SoX actualy
does in coreaudio.c when trying to open a device; which itself
is very much deprecated, starting with:

AudioHardwareGetProperty is deprecated: first deprecated in macOS 10.6
AudioHardwareGetPropertyInfo is deprecated: first deprecated in macOS 10.6

Tested on macOS 14.5, please test on all other versions.