build dependencies between files
This is a .depend file, base on mkdep(1) and edited manualy.
The actual dependencies still need to be verified file by file,
but this is good enough for the build.

The dependencies cen probably be loosened:
every audio format source, for example, dependes on sox_i.h,
and sox_i.h includes effects.h; so every format depends of effects.h.
But surely no audio format actualy depends on the decelration of
audio effects; similarly, no effect actualy depends on any other.

While here, rename to Makefile.depend
so that we don't miss it in 'grep ... *'