A first iteration of a soxformat(7) review
Rewrite the basic structures in mdoc(7)
and delineate the overall list of formats.
Some are fully rewritten into mdoc(7) already,
some are just cut out for later work.

While here, after nearly a decade, sync with reality a bit.
In particular, remove (in a separate commit) the formats
that the manpage of ten years ago says are "obsoleted".