get rid of the "decoration macros"
These are MSVC- and GCC-specific attributes telling the compiler
that the pointer returned by a function might be NULL (or not),
that the variables are initialized, that arguments are (un)used, etc.

We don't want to be telling GCC that;
we want $(CC) to compile our C code.

Also, we don't want to hide and ignore the fact that variables
might be uninitialized: to the contrary, we want $(CC) to tell us,
that's why we have -Wuninitialized in the CFLAGS.

Apparently, this was introduced in 2011 in commit
381a7bd0e2bbe4838cc59f9220eaecaa46b6a303 to

	override GCC's -mrtd switch
	or MSVC's -Gz or -Gr switches.
	Supported for MSVC and GCC.

and

	Functions that always return the same value
	(such as sox_get_globals()) are decorated as pure.
	Supported only by GCC.

This removes about ~500 lines of that.