xmalloc: fail or do what they wanted
The only point of these malloc wrappers is to
do the return check, so that every call doesn't have to.

So do just that: if it failed, it failed; otherwise,
return what the caller's argument(s) asked for,
including zero allocations.

In particular, don't pretend
that zero allocations are nonzero.