libsox(3): document sox_open_read() and the structs
After a decade, sync with reality.

This describes what sox_open_read() actualy is,
based on the actual source and the sox.h comments.
Remove those comments as we have actual documentation now.

Also document the structures passed to sox_open_read(),
because how else is the user supposed to know
how to specify e.g. the number of channels.
Straight from sox.h, including the typedef.

Obviously, this is just a start.
We need to do this with all libsox interfaces.